+++ src/providers/interswitch.provider.ts
@@ -1,4 +1,4 @@
-import { Providers, type BillPayCategory } from "../common";
+import { Providers, type BillaCategory } from "../common";
 import type { IBillPaymentProvider } from "../common/interfaces/bill-payment-provider";
 import type { BillerItem } from "../common/types/biller-item";
 import type { Customer, PayResponse } from "../common/types/payment";
@@ -31,7 +31,7 @@ export class InterswitchProvider implements IBillPaymentProvider {
     return this.interswitchService.getPlans(options);
   }
 
-  async listCategories(): Promise<BillPayCategory[]> {
+  async listCategories(): Promise<BillaCategory[]> {
     const res = await this.interswitchService.getBillerCategories();
     return (res.BillerCategories || []).map((cat) => ({
       name: cat.Name,